requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0

# Data processing and ML dependencies
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
keras==2.13.1
tensorflow==2.13.0

# Google Trends API
pytrends==4.9.0

# Model tracking and management
mlflow==2.7.1

# Configuration and environment
python-dotenv==1.0.0

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Development and debugging
Werkzeug==2.3.7 